:sparkles: Refactoring of project structure to delegate responsabilities
