+++ requirements.txt
@@ -1,3 +1,3 @@
 openai
 aiogram
-dotenv
+python-dotenv
