+++ src/bot/state.py
@@ -1,0 +1,2 @@
+# Dictionary to track user states
+user_states = {}
