+++ src/services/api.py
@@ -1,0 +1,14 @@
+import logging
+
+async def handle_api_transaction(api_response: dict):
+    """
+    Simulates handling an API transaction for financial records.
+    Replace this with actual API calls to save data.
+    """
+    note = api_response.get("note")
+    value = api_response.get("value")
+    type_ = api_response.get("type")
+
+    # Simulate saving to a database or external API
+    logging.info(f"API Transaction: Note: {note}, Value: {value}, Type: {type_}")
+    # Add actual API call logic here
